keyboard: re-add keyboard brightness handling
This is a squash of the following commits:

keyboard: re-add code to detect lid state

This was lost in the cm-11 merge. Without this, a change in lid state
isn't generated to light up the keyboard backlight

Author: Dave Daynard <nardholio@gmail.com>
Change-Id: Ief35cebcb62da13afe1ae7531d69ab58e196be9e

Keyboard light: Fix the physical keyboard not light when it's visible.

Author: YuanQY <yuanqingyun@gmail.com>
Change-Id: I595afd3cb6b422a17ae0f6ec20aa51979db13810

keyboard: adding functional alt/shift lights

This is useful for devices that has QWERTY keyboard
and leds for the Alt/Shift (Fn/Caps) keys, like
htc doubleshot.

Author: tbalden <illespal@gmail.com>

keyboard: remove special fn/caps handling

The special fn/caps light types need interface changes
in oreo, and there are only a handful of devices with
physical keyboards on nougat, and it's unlikely that any
of them will get ported to oreo.

Remove the special light types to avoid modifying the
hardware interfaces.

Author: Demon000 <demonsingur@gmail.com>

Change-Id: I66ebc2d881438f5b51db77eaa885421e65a7da0d